Unique items was falsy set for every types of collections
