+++ openapi-maven-plugin/src/test/resources/ut/JavadocParserTest/inheritance_test_three.yml
@@ -40,7 +40,6 @@ paths:
                     type: array
                     items:
                       type: string
-                    uniqueItems: true
                   hasNext:
                     description: True if a next slice exist
                     type: boolean
