+++ openapi-maven-plugin/src/test/resources/ut/SpringClassAnalyserTest/enumeration_test_1.yml
@@ -36,4 +36,3 @@ components:
           type: array
           items:
             $ref: '#/components/schemas/Authority'
-          uniqueItems: true
