+++ openapi-maven-plugin/src/test/resources/ut/SpringClassAnalyserTest/enumeration_test_1_with_default_errors.yml
@@ -50,7 +50,6 @@ components:
           type: array
           items:
             $ref: '#/components/schemas/Authority'
-          uniqueItems: true
     Error:
       type: object
       description: An error object
