+++ openapi-maven-plugin/src/test/resources/ut/SpringClassAnalyserTest/enumeration_test_4.yml
@@ -40,4 +40,3 @@ components:
           type: array
           items:
             $ref: '#/components/schemas/Authority'
-          uniqueItems: true
