+++ openapi-maven-plugin/src/test/resources/ut/SpringClassAnalyserTest/file_upload.yml
@@ -33,7 +33,6 @@ paths:
                   items:
                     type: string
                     format: binary
-                  uniqueItems: true
       responses:
         200:
           description: successful operation
@@ -66,7 +65,6 @@ paths:
                   items:
                     type: string
                     format: binary
-                  uniqueItems: true
       responses:
         200:
           description: successful operation
