+++ openapi-maven-plugin/src/test/resources/ut/SpringClassAnalyserTest/json_ignore.yml
@@ -54,7 +54,6 @@ components:
               type: array
               items:
                 $ref: '#/components/schemas/Authority'
-              uniqueItems: true
             hasNext:
               type: boolean
             totalPages:
