+++ openapi-maven-plugin/src/test/resources/ut/SpringClassAnalyserTest/nested_genericity.yml
@@ -57,7 +57,6 @@ paths:
                       type: array
                       items:
                         $ref: '#/components/schemas/AccountDto'
-                      uniqueItems: true
                     hasNext:
                       type: boolean
     post:
@@ -74,7 +73,6 @@ paths:
                   type: array
                   items:
                     $ref: '#/components/schemas/AccountDto'
-                  uniqueItems: true
                 hasNext:
                   type: boolean
                 totalPages:
@@ -107,7 +105,6 @@ paths:
                     type: array
                     items:
                       $ref: '#/components/schemas/Authority'
-                    uniqueItems: true
                   hasNext:
                     type: boolean
                   totalPages:
@@ -164,7 +161,6 @@ paths:
                     type: array
                     items:
                       $ref: '#/components/schemas/TimeDto'
-                    uniqueItems: true
                   hasNext:
                     type: boolean
 components:
